9주차 과제 Follow 등록, 푸시알림 추가

+++ src/main/java/com/springboot/domain/category/entity/CategoryRepository.java
@@ -11,6 +11,6 @@ public interface CategoryRepository extends JpaRepository<Category, Long> {
 
     List<Category> findAllByMemberId(Long memberId);
 
-    Optional<Category> findByMemberIdAndCategoryCode(BigInteger memberId, CategoryCode categoryCode);
+    Optional<Category> findByMemberIdAndCategoryCode(Long memberId, CategoryCode categoryCode);
 
 }
